Fromm 2/10 testing updating Autonomous, Dispenser, Drive, PidGyro, Robot166, & Vision166.cpp.

+++ chopshop09/Autonomous166.h
@@ -27,7 +27,7 @@ class autonomous166
         float archived_distance;             // the current distance to the target
   		float drop_distance;                 // distance at which the dropper can be activated
   		float activate_drop;                 // activates dropping function variable
-  		float gyrate;                        // activates gyrating action for the motor 
+  		int gyrate;                        // activates gyrating action for the motor 
 	
 	
 };
